Add instructions to run tests
